Update to NeRFICG v2.0 (adds Windows support)
Apart from adding Windows support and updating setup instructions, this commit fixes a few typos and includes some refactoring. Results remain consistent with the initial release.

+++ INPCCudaBackend/INPCCudaBackend/sampling/include/sampling_utils.h
@@ -14,7 +14,7 @@ namespace inpc::sampling {
         int* sample_counts = nullptr;
         int* sample_offsets = nullptr;
 
-        static SumBuffers from_blob(char*& blob, const size_t n_cells) {
+        static SumBuffers from_blob(char*& blob, const uint n_cells) {
             SumBuffers buffers;
             float* float_ptr = nullptr;
             size_t reduce_sum_size = 0;
